working towards support for SAMRAI 3.9.0

+++ ibtk/src/lagrangian/LTransaction.cpp
@@ -71,7 +71,7 @@ LTransaction<T>::LTransaction(const int src_proc,
                               const std::vector<LTransactionComponent>& src_item_set)
     : d_src_item_set(src_item_set), d_src_proc(src_proc), d_outgoing_bytes(0), d_dst_item_set(), d_dst_proc(dst_proc)
 {
-    d_outgoing_bytes = MessageStream::getSizeof<int>();
+    d_outgoing_bytes = static_cast<int>(MessageStream::getSizeof<int>());
     for (auto cit = d_src_item_set.begin(); cit != d_src_item_set.end(); ++cit)
     {
         d_outgoing_bytes += cit->item->getDataStreamSize() + NDIM * MessageStream::getSizeof<double>();
